requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai>=1.12.0
pillow==10.1.0
PyMuPDF>=1.23.0
pytesseract==0.3.10
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
numpy>=1.26.0
opencv-python==4.8.1.78
setuptools>=65.0.0

